fix compiler issues
