+++ packages/common-test-utils/src/index.ts
@@ -1,10 +1,9 @@
 import {
   DNodeUtils,
   DVault,
-  NoteChangeEntry,
   NoteOpts,
-  NotePropsDict,
   NoteProps,
+  NotePropsDict,
   NoteUtils,
   SchemaModuleOpts,
   SchemaModuleProps,
@@ -21,12 +20,11 @@ import {
   tmpDir,
   vault2Path,
 } from "@dendronhq/common-server";
-import assert from "assert";
 import fs from "fs-extra";
 import _ from "lodash";
 import path from "path";
-import { PreSetupHookFunctionV4, TestResult } from "./types";
-import { AssertUtils, TestPresetEntry } from "./utils";
+import { PreSetupHookFunctionV4 } from "./types";
+import { AssertUtils } from "./utils";
 export * from "./fileUtils";
 export * from "./noteUtils";
 export * from "./presets";
@@ -460,353 +458,5 @@ export class NodeTestUtilsV2 {
       return NodeTestUtilsV2.normalizeNote({ note });
       //return { ..._.omit(note, ["body", "parent", "id"]), body: _.trim(note.body) };
     });
-  }
-}
-
-export class NodeTestPresetsV2 {
-  static async runMochaHarness<TOpts>({
-    opts,
-    results,
-  }: {
-    opts: TOpts;
-    results: any;
-  }) {
-    return _.map(await results(opts), (ent) =>
-      assert.deepStrictEqual(ent.actual, ent.expected)
-    );
-  }
-  static async runJestHarness<TOpts>({
-    opts,
-    results,
-    expect,
-  }: {
-    opts: TOpts;
-    results: Function;
-    expect: jest.Expect;
-  }) {
-    return _.map(await results(opts), (ent) =>
-      expect(ent.actual).toEqual(ent.expected)
-    );
-  }
-
-  static async createNoteRefPreset({ vaultDir }: { vaultDir: string }) {
-    await NodeTestPresetsV2.createOneNoteOneSchemaPresetWithBody({ vaultDir });
-    await NodeTestUtilsV2.createNote({
-      vaultDir,
-      noteProps: {
-        body: "![[foo]]",
-        fname: "bar",
-        vault: { fsPath: vaultDir },
-      },
-    });
-  }
-
-  static async createSchemaPreset({ vaultDir }: { vaultDir: string }) {
-    const vault = { fsPath: vaultDir };
-    await NodeTestPresetsV2.createOneNoteOneSchemaPreset({ vaultDir });
-    await NodeTestUtilsV2.createSchema({
-      vaultDir,
-      schemas: [
-        SchemaUtils.create({
-          id: "bar",
-          parent: "root",
-          children: ["ch1", "ch2"],
-          vault,
-        }),
-        SchemaUtils.create({
-          id: "ch1",
-          template: { id: "bar.template.ch1", type: "note" },
-          vault,
-        }),
-        SchemaUtils.create({
-          id: "ch2",
-          template: { id: "bar.template.ch2", type: "note" },
-          namespace: true,
-          vault,
-        }),
-      ],
-      fname: "bar",
-    });
-    await NodeTestUtilsV2.createNote({
-      vaultDir,
-      noteProps: { body: "ch1 template", fname: "bar.template.ch1", vault },
-    });
-    await NodeTestUtilsV2.createNote({
-      vaultDir,
-      noteProps: {
-        body: "ch2 template",
-        fname: "bar.template.ch2",
-        vault,
-      },
-    });
-  }
-
-  static async createOneNoteOneSchemaPresetWithBody({
-    vaultDir,
-  }: {
-    vaultDir: string;
-  }) {
-    await NodeTestUtilsV2.createSchemas({ vaultPath: vaultDir });
-    await NodeTestUtilsV2.createNotes({
-      vaultPath: vaultDir,
-    });
-    await NodeTestUtilsV2.createNoteProps({
-      vaultPath: vaultDir,
-      rootName: "foo",
-      props: {
-        body: "foo body",
-      },
-    });
-    await NodeTestUtilsV2.createSchemaModuleOpts({
-      vaultDir: vaultDir,
-      rootName: "foo",
-    });
-  }
-  static async createOneNoteOneSchemaPreset({
-    vaultDir,
-  }: {
-    vaultDir: string;
-  }) {
-    await NodeTestUtilsV2.createSchemas({ vaultPath: vaultDir });
-    await NodeTestUtilsV2.createNotes({ vaultPath: vaultDir });
-    await NodeTestUtilsV2.createNoteProps({
-      vaultPath: vaultDir,
-      rootName: "foo",
-    });
-    await NodeTestUtilsV2.createSchemaModuleOpts({
-      vaultDir: vaultDir,
-      rootName: "foo",
-    });
-  }
-}
-
-type NoteTestPresetGroup<TBeforeOpts, TResultsOpts> = {
-  [key: string]: TestPresetEntry<TBeforeOpts, any, TResultsOpts>;
-};
-
-type NoteTestPresetModule<TBeforeOpts, TResultsOpts> = {
-  [key: string]: NoteTestPresetGroup<TBeforeOpts, TResultsOpts>;
-};
-
-type NoteTestPresetCollectionDict<TBeforeOpts = any, TResultsOpts = any> = {
-  [key: string]: NoteTestPresetModule<TBeforeOpts, TResultsOpts>;
-};
-
-type DeleteNoteTestOptsV2 = {
-  changed: NoteChangeEntry[];
-  notes: NotePropsDict;
-  vaultDir: string;
-};
-
-type UpdateNoteTestOptsV2 = {
-  notes: NotePropsDict;
-  vaultDir: string;
-};
-
-export class NoteTestPresetsV2 {
-  static async createJestTest({
-    executeCb,
-    beforeArgs,
-    expect,
-    entry,
-  }: {
-    entry: TestPresetEntry<any, any, any>;
-    expect: jest.Expect;
-    beforeArgs?: any;
-    executeCb: () => Promise<any>;
-  }) {
-    await entry.before(beforeArgs);
-    const results = entry.results;
-    const executeResp = await executeCb();
-    await NodeTestPresetsV2.runJestHarness({
-      opts: executeResp,
-      results,
-      expect,
-    });
-  }
-
-  static presets: NoteTestPresetCollectionDict = {
-    OneNoteOneSchemaPreset: {
-      init: {
-        domainStub: new TestPresetEntry({
-          label: "domain stub",
-          before: async ({ vaultDir }: { vaultDir: string }) => {
-            fs.removeSync(path.join(vaultDir, "foo.md"));
-          },
-          results: async ({ notes }: { notes: NotePropsDict }) => {
-            const note = NoteUtils.getNoteByFname("foo", notes) as NoteProps;
-            const vault = note.vault;
-            const root = NoteUtils.getNoteByFname("root", notes, {
-              vault,
-            }) as NoteProps;
-            const scenarios = [
-              { actual: _.size(notes), expected: 3 },
-              { actual: root.children, expected: [note.id] },
-            ];
-            return scenarios;
-          },
-        }),
-      },
-      delete: {
-        noteNoChildren: new TestPresetEntry({
-          label: "note w/no children",
-          results: NoteTestPresetsV2.createDeleteNoteWNoChildrenResults,
-        }),
-        domainChildren: new TestPresetEntry({
-          label: "domain with children",
-          results: async ({
-            changed,
-            notes,
-            vaultDir,
-          }: DeleteNoteTestOptsV2): Promise<TestResult[]> => {
-            return [
-              {
-                actual: changed,
-                expected: [{ note: notes["foo"], status: "update" }],
-                msg: "note updated",
-              },
-              {
-                actual: _.size(notes),
-                expected: 3,
-                msg: "same number of notes",
-              },
-              {
-                actual: notes["foo"].stub,
-                expected: true,
-                msg: "foo should be a stub",
-              },
-              {
-                actual: _.includes(fs.readdirSync(vaultDir), "foo.md"),
-                expected: false,
-                msg: "note should be deleted",
-              },
-            ];
-          },
-        }),
-        domainNoChildren: new TestPresetEntry({
-          label: "domain w/no children",
-          results: async ({
-            changed,
-            notes,
-            vaultDir,
-          }: DeleteNoteTestOptsV2): Promise<TestResult[]> => {
-            return [
-              {
-                actual: changed[0].note.fname,
-                expected: "root",
-                msg: "root updated",
-              },
-              {
-                actual: changed[0].note.children,
-                expected: [],
-                msg: "root does not have children",
-              },
-              { actual: _.size(notes), expected: 1 },
-              { actual: notes["foo"], expected: undefined },
-              {
-                actual: _.includes(fs.readdirSync(vaultDir), "foo.md"),
-                expected: false,
-              },
-            ];
-          },
-        }),
-      },
-      update: {
-        noteNoChildren: new TestPresetEntry({
-          label: "update note, no children",
-          results: async ({
-            notes,
-          }: UpdateNoteTestOptsV2): Promise<TestResult[]> => {
-            return [
-              {
-                actual: _.pick(notes["foo.ch1"], "body"),
-                expected: { body: "new body" },
-                msg: "update body",
-              },
-            ];
-          },
-        }),
-      },
-      write: {
-        domainStub: new TestPresetEntry({
-          label: "write child, parent stub",
-          before: async ({ vaultDir }: { vaultDir: string }) => {
-            const note = NoteUtils.create({
-              fname: "bar.ch1",
-              vault: { fsPath: vaultDir },
-            });
-            await note2File({
-              note,
-              vault: { fsPath: vaultDir },
-              wsRoot: "FAKE",
-            });
-          },
-          results: async ({ notes }: { notes: NotePropsDict }) => {
-            const root = NoteUtils.getNoteByFname("root", notes) as NoteProps;
-            const bar = NoteUtils.getNoteByFname("bar", notes) as NoteProps;
-            const child = NoteUtils.getNoteByFname(
-              "bar.ch1",
-              notes
-            ) as NoteProps;
-            return [
-              {
-                actual: _.size(root.children),
-                expected: 2,
-                msg: "root, foo, bar",
-              },
-              {
-                actual: _.pick(bar, "stub"),
-                expected: { stub: true },
-                msg: "bar created as stub",
-              },
-              {
-                actual: _.pick(child, ["fname", "stub"]),
-                expected: { fname: "bar.ch1" },
-                msg: "child is not stub",
-              },
-            ];
-          },
-        }),
-        serializeChildWithHierarchy: new TestPresetEntry({
-          label: "write child, serialize with hierarchy",
-          before: async ({ vaultDir }: { vaultDir: string }) => {
-            fs.removeSync(path.join(vaultDir, "foo.ch1.md"));
-          },
-          results: async ({
-            vaultDir,
-          }: {
-            notes: NotePropsDict;
-            vaultDir: string;
-          }) => {
-            const rawNote = fs.readFileSync(path.join(vaultDir, "foo.ch1.md"), {
-              encoding: "utf8",
-            });
-            return [
-              {
-                actual: rawNote.match(/^parent: foo/gm),
-                expected: ["parent: foo"],
-                msg: "should have parent",
-              },
-            ];
-          },
-        }),
-      },
-    },
-  };
-
-  static async createDeleteNoteWNoChildrenResults({
-    changed,
-    notes,
-    vaultDir,
-  }: DeleteNoteTestOptsV2): Promise<TestResult[]> {
-    return [
-      { actual: changed[0].note.id, expected: "foo" },
-      { actual: _.size(notes), expected: 2 },
-      { actual: notes["foo"].children, expected: [] },
-      {
-        actual: _.includes(fs.readdirSync(vaultDir), "foo.ch1.md"),
-        expected: false,
-      },
-    ];
   }
 }
