+++ packages/common-test-utils/src/utilsv2.ts
@@ -12,7 +12,6 @@ import {
   SetupTestFunctionV4,
 } from ".";
 import { PostSetupHookFunction, PreSetupHookFunction } from "./types";
-import { Node as UnistNode } from "unist";
 
 type EngineOverride = {
   [P in keyof DEngineClient]: (opts: WorkspaceOpts) => DEngineClient[P];
@@ -91,23 +90,4 @@ export class TestPresetEntryV4 {
       },
     ];
   }
-}
-
-/** Gets the descendent (child, or child of child...) node of a given node.
- *
- * @param node The root node to start descending from.
- * @param indices Left-to-right indexes for children, e.g. first index is for the root, second is for the child of the root...
- * @returns Requested child. Note that this function has no way of checking types, so the child you get might not be of the right type.
- */
-export function getDescendantNode<Child extends UnistNode>(
-  node: UnistNode,
-  ...indices: number[]
-): Child {
-  const index = indices.shift();
-  if (_.isUndefined(index)) return node as Child;
-  expect(node).toHaveProperty("children");
-  expect(node.children).toHaveProperty("length");
-  const children = node.children as UnistNode[];
-  expect(children.length).toBeGreaterThanOrEqual(index);
-  return getDescendantNode<Child>(children[index], ...indices);
 }
