+++ packages/engine-test-utils/src/__tests__/engine-server/markdown/blockAnchors.spec.ts
@@ -1,8 +1,4 @@
-import {
-  AssertUtils,
-  TestPresetEntryV4,
-  getDescendantNode,
-} from "@dendronhq/common-test-utils";
+import { AssertUtils, TestPresetEntryV4 } from "@dendronhq/common-test-utils";
 import {
   BlockAnchor,
   DendronASTDest,
@@ -16,7 +12,10 @@ import {
 import _ from "lodash";
 import { runEngineTestV5 } from "../../../engine";
 import { ENGINE_HOOKS } from "../../../presets";
+import { TestUnifiedUtils } from "../../../utils";
 import { createProcForTest, createProcTests, ProcTests } from "./utils";
+
+const { getDescendantNode } = TestUnifiedUtils;
 
 function proc() {
   return MDUtilsV5.procRehypeParse({
