+++ packages/engine-test-utils/src/presets/pods-core/index.ts
@@ -1,9 +1,7 @@
-import SNAPSHOT from "./snapshot";
 import JSON from "./json";
 import MARKDOWN from "./markdown";
 
 export const PODS_CORE = {
-  SNAPSHOT,
   JSON,
   MARKDOWN,
 };
